Split expectEvent testing
